# backend/openapi.yaml
openapi: 3.1.0
info:
  title: GUTTERCAPS API
  version: 0.2.0
  description: |
    Read-mostly API for the GUTTERCAPS client. Anything that moves value is a
    Solana transaction the client builds and signs itself; the API only
    supplies quotes, PDAs, Merkle proofs and projections.

    Auth: Sign-In-With-Solana (SIWS). `POST /auth/siws/nonce` → sign → `POST /auth/siws/verify` → HTTP-only cookie + CSRF token.
    The signed message's domain must be one of the API's `SIWS_DOMAINS` and carry `Issued At` within ±5 min.
    Rate limits (every endpoint may answer `429 rate_limited` with `Retry-After` and `RateLimit-Limit/Remaining/Reset/Policy`
    headers): reads 600/min/IP · mutations 60/min/session · `/auth/siws/nonce` 10/min/IP + 30/h/wallet · `/auth/siws/verify`
    20/min/IP · `/packs/quote` 30/min/session · `/services/claim`, `PUT /me/handle` 10/min/session + 40/min per IP /24 ·
    `POST /me/human` 6/min/session + 30/h per IP /24. JSON bodies ≤ 16 KB (`413`).
    All amounts are strings of integer base units (lamports / micro-USDC / micro-$CG) unless suffixed `Usd`.
servers:
  - url: https://api.guttercaps.gg/v1
  - url: http://localhost:8787/v1
...
security:
  - cookieAuth: []
paths:
...
  /auth/siws/nonce:
    post:
      tags: [auth]
      security: []
      summary: Issue a SIWS nonce for a wallet
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [address], properties: { address: { $ref: '#/components/schemas/Pubkey' } } } } }
      responses:
        '200': { description: nonce, content: { application/json: { schema: { type: object, properties: { nonce: { type: string }, statement: { type: string }, expiresAt: { type: string, format: date-time } } } } } }
  /auth/siws/verify:
    post:
      tags: [auth]
      security: []
      summary: Verify signed SIWS message, start session
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [address, message, signature], properties: { address: { $ref: '#/components/schemas/Pubkey' }, message: { type: string }, signature: { type: string, description: base58 }, referrer: { $ref: '#/components/schemas/Pubkey' }, fingerprint: { type: string, minLength: 8, maxLength: 256, description: 'client device fingerprint (canvas-free, client/src/shared/lib/fingerprint.ts); the API stores only a salted hash — device dedupe, T-B-49' } } } } }
      responses:
        '200': { description: session, content: { application/json: { schema: { type: object, properties: { csrf: { type: string }, wallet: { $ref: '#/components/schemas/WalletProfile' } } } } } }
        '401': { description: 'siws_address | siws_nonce | siws_expired | siws_domain | siws_issued_at | siws_signature', content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } } }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /me/chips:
    get:
      tags: [me]
      summary: Inventory (projection of Core owner + ChipState)
      parameters:
        - { in: query, name: collection, schema: { type: integer, minimum: 0, maximum: 9 } }
        - { in: query, name: rarity, schema: { type: integer, minimum: 0, maximum: 8 } }
        - { in: query, name: status, schema: { type: string, enum: [free, staked, listed, locked, fusing] } }
        - { $ref: '#/components/parameters/cursor' }
      responses:
        '200': { description: chips, content: { application/json: { schema: { $ref: '#/components/schemas/ChipPage' } } } }
...
  /me/activity:
    get:
      tags: [me]
      summary: Unified activity feed (pack opens, fusions, sales, claims, matches)
      parameters: [{ $ref: '#/components/parameters/cursor' }]
      responses: { '200': { description: feed, content: { application/json: { schema: { $ref: '#/components/schemas/ActivityPage' } } } } }
...
  /me/handle/check:
    get:
      tags: [me]
      summary: Is this handle free? (also returns the ref_hash to commit on-chain)
      description: |
        ref_hash = keccak256(0x00 ‖ kind:u8 ‖ wallet:32 ‖ lowercase(handle) utf8). The client passes it to
        chip_core::pay_service(kind=Handle|HandleChange, …). Reserved 120 s for the caller once checked.
      parameters: [{ in: query, name: handle, required: true, schema: { type: string, pattern: '^[a-zA-Z0-9_]{3,16}$' } }]
      responses:
        '200': { description: availability, content: { application/json: { schema: { type: object, properties: { available: { type: boolean }, reason: { type: string, enum: [taken, reserved, blocked, cooldown] }, kind: { type: integer, description: '0 first handle, 1 change' }, refHash: { type: string, description: 'hex, 32 bytes' }, priceUsdCents: { type: integer }, reservedUntil: { type: string, format: date-time } } } } } }
  /me/handle:
    put:
      tags: [me]
      summary: Claim a handle after paying on-chain (pay_service kind 0/1) — signature required
      description: |
        The indexer verifies the ServicePaid event in `signature`: buyer == caller, kind ∈ {0,1}, ref_hash matches
        keccak(kind ‖ wallet ‖ lowercase(handle)), payment not previously consumed. $CG payments are burned on-chain;
        SOL/USDC/SKR go to the treasury. Handle changes are limited to 1 per 30 days; released handles are quarantined 90 days.
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [handle, signature], properties: { handle: { type: string, pattern: '^[a-zA-Z0-9_]{3,16}$' }, signature: { type: string } } } } } }
      responses: { '200': { description: ok }, '402': { description: payment not found / already consumed }, '409': { $ref: '#/components/responses/Error', description: 'handle_taken / handle_cooldown / same_handle — or payment_pending: the payment is confirmed but not yet finalized (SEC-M5); retry in ~30 s' } }
...
  /me/human:
    get:
      tags: [me]
      summary: Proof-of-human state (Cloudflare Turnstile pass, T-B-49)
      description: |
        Quest and SKR rewards are only *settled* for wallets with a fresh pass (7 d); until then `/quests` items carry
        `ineligibleReason = human_check_required` and finished quests wait (nothing is zeroed). `required=false` when the
        deployment runs without Turnstile or the wallet is `trusted`. `siteKey` is what the widget needs.
      responses:
        '200': { description: state, content: { application/json: { schema: { $ref: '#/components/schemas/HumanStatus' } } } }
    post:
      tags: [me]
      summary: Verify a Turnstile token → 7-day pass
      description: 'Rate limits: 6/min per session, 30/h per IP /24. Optionally re-sends the device fingerprint.'
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [token], properties: { token: { type: string, maxLength: 2048 }, fingerprint: { type: string, minLength: 8, maxLength: 256 } } } } } }
      responses:
        '200': { description: verified, content: { application/json: { schema: { $ref: '#/components/schemas/HumanStatus' } } } }
        '400': { $ref: '#/components/responses/Error', description: 'turnstile_failed (details.errorCodes = siteverify error-codes)' }
        '503': { $ref: '#/components/responses/Error', description: 'human_check_unavailable — siteverify unreachable, retry' }
...
  /services/claim:
    post:
      tags: [services]
      summary: Bind a paid ServicePaid event to an entitlement payload (skins → asset, theme id, emote pack id, …)
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [signature, kind, payload], properties: { signature: { type: string }, kind: { type: integer }, payload: { type: object, additionalProperties: true, description: 'canonical JSON; keccak(kind ‖ wallet ‖ canonical(payload)) must equal ref_hash' } } } } } }
      responses: { '200': { description: entitlement, content: { application/json: { schema: { $ref: '#/components/schemas/Entitlement' } } } }, '402': { description: payment not found / mismatch }, '409': { $ref: '#/components/responses/Error', description: 'not_owner — or payment_pending: the payment is confirmed but not yet finalized (SEC-M5); retry in ~30 s' } }

  # ------------------------------------------------------------------ packs
...
  /packs/quote:
    post:
      tags: [packs]
      summary: Price quote + everything needed to build buy_pack
      description: |
        Returns the SOL / SKR amount computed from the studio's own Pyth push-oracle accounts (shard 0xCA75, posted by
        our price pusher — ops/pyth-pusher/) with the exact integer formula of chip_core::units_for_cents, the 1 %
        slippage guard (`maxLamports`), the `priceUpdateAccount` the client must pass to buy_pack, $CG/USDC amounts,
        the caller's effective odds (pity applied), the Switchboard queue and the PDA set. Bundle and SKR promo
        discounts are stacked additively and capped at 30 %.
        The quote is never synthesised: if the on-chain price is older than 45 s (60 s max age minus the 15 s a buyer
        needs to sign) or the account is missing/unverified the endpoint answers **503 price_unavailable** and the
        client hides that rail. `expiresAt` is when the current on-chain update would fail the age check.
        `pythUpdateData` is always empty in the push model (kept for a future pull-model fallback).
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [sku, qty, currency], properties: { sku: { type: integer, minimum: 0, maximum: 3 }, qty: { type: integer, minimum: 1, maximum: 25 }, currency: { type: string, enum: [SOL, USDC, CG, SKR] } } } } } }
      responses:
        '200': { description: quote, content: { application/json: { schema: { $ref: '#/components/schemas/PackQuote' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '401': { description: sign-in required (pity and caps are per wallet) }
        '409': { description: starter_claimed — one Starter per wallet }
        '429': { description: daily_cap — Limited 5 / day }
        '503': { $ref: '#/components/responses/Error', description: 'price_unavailable — Pyth account stale / missing / unverified, or its confidence interval is wider than 2 % of the price (SEC-M2 — the program would reject with PriceUncertain); retry after the next push (≤ 30 s)' }
...
  /packs/opens/{signature}:
    get:
      tags: [packs]
      summary: Result of a pack open (poll after reveal; WS pushes the same payload)
      parameters: [{ in: path, name: signature, required: true, schema: { type: string } }]
      responses:
        '200': { description: result, content: { application/json: { schema: { $ref: '#/components/schemas/PackOpenResult' } } } }
        '202': { description: not revealed yet, content: { application/json: { schema: { $ref: '#/components/schemas/PendingPack' } } } }
  /packs/verify:
    post:
      tags: [packs]
      security: []
      summary: Provably-fair verifier — recompute a pack result from its randomness bytes
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [signature], properties: { signature: { type: string } } } } } }
      responses:
        '200': { description: verification, content: { application/json: { schema: { type: object, properties: { randomnessAccount: { $ref: '#/components/schemas/Pubkey' }, rollHex: { type: string }, pityBefore: { type: integer }, effectiveOddsBps: { type: array, items: { type: integer }, description: 'odds the roll was drawn with — the SKU table after pity, or the voucher TEMPLATE odds (#28)' }, voucher: { type: [object, 'null'], properties: { template: { type: integer }, odds: { type: [array, 'null'], items: { type: integer } }, soulboundDays: { type: [integer, 'null'] } } }, recomputed: { type: array, items: { $ref: '#/components/schemas/RolledChip' } }, onChain: { type: array, items: { $ref: '#/components/schemas/RolledChip' } }, matches: { type: boolean } } } } } }

  # ------------------------------------------------------------------ collections
...
  /chips/{asset}:
    get:
      tags: [collections]
      security: []
      summary: Single chip — state, owner, listing, provenance (pack/fusion event)
      parameters: [{ in: path, name: asset, required: true, schema: { $ref: '#/components/schemas/Pubkey' } }]
      responses: { '200': { description: chip, content: { application/json: { schema: { $ref: '#/components/schemas/ChipDetail' } } } }, '404': { $ref: '#/components/responses/Error' } }

  # ------------------------------------------------------------------ market
  /market/listings:
    get:
      tags: [market]
      security: []
      summary: Active listings with filters
      parameters:
        - { in: query, name: collection, schema: { type: integer } }
        - { in: query, name: rarity, schema: { type: integer } }
        - { in: query, name: rarityMin, schema: { type: integer } }
        - { in: query, name: indexMin, schema: { type: integer } }
        - { in: query, name: indexMax, schema: { type: integer } }
        - { in: query, name: levelMin, schema: { type: integer } }
        - { in: query, name: currency, schema: { type: string, enum: [SOL, USDC, SKR] } }
        - { in: query, name: priceMaxUsd, schema: { type: number } }
        - { in: query, name: missingForMySet, schema: { type: boolean }, description: 'auth required; only chips that complete one of my sets' }
        - { in: query, name: sort, schema: { type: string, enum: [price_asc, price_desc, newest, rarity_desc, index_asc] } }
        - { $ref: '#/components/parameters/cursor' }
      responses: { '200': { description: listings, content: { application/json: { schema: { $ref: '#/components/schemas/ListingPage' } } } } }
...
  /market/history:
    get:
      tags: [market]
      security: []
      summary: Sales history (per chip, archetype, or global)
      parameters:
        - { in: query, name: asset, schema: { $ref: '#/components/schemas/Pubkey' } }
        - { in: query, name: collection, schema: { type: integer } }
        - { in: query, name: rarity, schema: { type: integer } }
        - { $ref: '#/components/parameters/cursor' }
      responses: { '200': { description: sales, content: { application/json: { schema: { $ref: '#/components/schemas/SalePage' } } } } }
  /market/offers:
    get:
      tags: [market]
      summary: Offers made by me / received on my chips
      parameters: [{ in: query, name: direction, schema: { type: string, enum: [made, received] } }]
      responses: { '200': { description: offers, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Offer' } } } } } }

  # ------------------------------------------------------------------ fusion
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: gc_session }
  parameters:
    cursor: { in: query, name: cursor, schema: { type: string }, description: opaque pagination cursor }
  responses:
    Error:
      description: error
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
...
  schemas:
    ApiError: { type: object, properties: { code: { type: string }, message: { type: string }, details: {} } }
    Pubkey: { type: string, pattern: '^[1-9A-HJ-NP-Za-km-z]{32,44}$' }
...
    WalletProfile:
      type: object
      properties: { address: { $ref: '#/components/schemas/Pubkey' }, handle: { type: string }, avatarAsset: { $ref: '#/components/schemas/Pubkey' }, firstSeen: { type: string, format: date-time }, country: { type: string } }
...